requirements.txt
numpy
Pillow
bcrypt
PyQt5
pypinyin
openpyxl